Revert "refactor: Renamed JsonArrayLoader to UrlArrayLoader"
This reverts commit e6396ef3983f77c42ef32e0a887d2de8710e8cd9.
