+++ src/colorizer/loaders/JsonArrayLoader.ts
